This function prints all nodes in a single list

+++ 0x12-singly_linked_lists/list.h
@@ -6,17 +6,17 @@
 #include <string.h>
 
 /**
- * struct list_s - these are singly linked lists
+ * struct list_s - singly linked list
+ * @str: string - (malloc'ed string)
+ * @len: length of the string
+ * @next: points to the next node
  *
- * @str: The malloc string
- * @len: The length of the string
- * @next: The pointer to the next node
+ * Description: singly linked list node structure
  */
-
 typedef struct list_s
 {
 	char *str;
-	size_t len;
+	unsigned int len;
 	struct list_s *next;
 } list_t;
